+++ xLights/outputs/DDPOutput.h
@@ -63,9 +63,9 @@ public:
     void SetChannelsPerPacket(int cpp) { _channelsPerPacket = cpp; _dirty = true; }
     virtual std::string GetType() const override { return OUTPUT_DDP; }
     virtual std::string GetLongDescription() const override;
-    virtual std::string GetChannelMapping(long ch) const override;
+    virtual std::string GetChannelMapping(int32_t ch) const override;
     virtual int GetMaxChannels() const override { return 1000000; }
-    virtual bool IsValidChannelCount(long channelCount) const override { return channelCount > 0 && channelCount <= GetMaxChannels(); }
+    virtual bool IsValidChannelCount(int32_t channelCount) const override { return channelCount > 0 && channelCount <= GetMaxChannels(); }
     virtual bool IsKeepChannelNumbers() const { return _keepChannelNumbers; }
     virtual void KeepChannelNumber(bool b = true) { _keepChannelNumbers = b; _dirty = true; }
     virtual bool IsLookedUpByControllerName() const override;
@@ -88,8 +88,8 @@ public:
     #pragma endregion Frame Handling
 
     #pragma region Data Setting
-    virtual void SetOneChannel(long channel, unsigned char data) override;
-    virtual void SetManyChannels(long channel, unsigned char* data, long size) override;
+    virtual void SetOneChannel(int32_t channel, unsigned char data) override;
+    virtual void SetManyChannels(int32_t channel, unsigned char* data, size_t size) override;
     virtual void AllOff() override;
     #pragma endregion Data Setting
